ViteJS config file with postcssrc file. Removed all webpack config files. Updated package.json with lock file
