+++ package.json
@@ -20,8 +20,9 @@
   ],
   "scripts": {
     "test": "jest --passWithNoTests",
-    "dev": "export NODE_ENV=development && webpack serve --hot --config=./webpack.config.js",
-    "build": "export NODE_ENV=production && webpack --config=./webpack.config.js --mode=production --bail",
+    "dev": "export NODE_ENV=development && vite --config vite.config.js",
+    "build": "vite build",
+    "preview": "vite preview",
     "prettier": "prettier ./src/js/*/*.js ./src/scss/**/*.scss --write",
     "modules-cli": "node ./_modules-cli/cli.js"
   },
@@ -30,36 +31,20 @@
   },
   "homepage": "https://github.com/Mill3/mill3-wp-theme-boilerplate",
   "devDependencies": {
-    "@swc/core": "^1.3.95",
-    "assets-webpack-plugin": "^7.1.1",
-    "autoprefixer": "^10.4.13",
+    "@fullhuman/postcss-purgecss": "^7.0.2",
     "chalk": "4.1.2",
-    "css-loader": "^6.7.3",
-    "css-minimizer-webpack-plugin": "^4.2.2",
-    "debounce": "^1.2.1",
     "dotenv": "^16.0.3",
     "ejs": "^3.1.8",
     "fs-extra": "^11.1.0",
-    "husky": "^3.0.9",
     "jest": "^27.0.6",
-    "mini-css-extract-plugin": "^2.7.2",
-    "path": "^0.12.7",
-    "postcss": "^8.4.21",
-    "postcss-loader": "^7.0.2",
-    "postcss-preset-env": "^7.8.3",
-    "postcss-scss": "^4.0.6",
-    "sass": "^1.57.1",
-    "sass-loader": "^13.2.0",
-    "style-loader": "^3.3.1",
-    "swc-loader": "^0.2.3",
-    "webpack": "^5.89.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1",
-    "webpackbar": "^5.0.2",
+    "postcss-preset-env": "^10.1.5",
+    "postcss-scss": "^4.0.9",
+    "sass-embedded": "^1.86.0",
+    "vite": "^6.2.3",
     "yargs-parser": "^21.1.1"
   },
   "dependencies": {
-    "@rive-app/canvas-lite": "^2.25.2",
+    "@rive-app/canvas": "^2.27.0",
     "@sentry/browser": "8.25.0",
     "@vimeo/player": "2.23.0",
     "animejs": "3.2.2",
